requirements.txt
allure-pytest==2.13.5
pytest==8.3.3
pytest-mock==3.14.0
selenium==4.27.1
webdriver-manager==4.0.2

